// interfaces/lcu-types.d.ts
export interface LolTftTrovesTrovesBanner {
	id: string
	/** @format uint8 */
	version: number
	sourceId: string
	startDate: string
	endDate: string
	/** @format uint32 */
	pityLimit: number
	rollOffer: string
	mythicOffer: string
	bannerTexture: string
	thumbnailTexture: string
	backgroundTexture: string
	platformTexture: string
	eventHubBannerTexture: string
	name: string
	description: string
	pityCounterId: string
	isCollectorBounty: boolean
	/** @format uint32 */
	maxTotalRolls: number
	/** @format uint32 */
	pullCost: number
	chaseContentId: string
	celebrationTheme: LolTftTrovesTrovesCelebrationThemeData
	status: LolTftTrovesTrovesStatus
}

export interface LolTftTrovesTrovesCelebrationCurrencySegmentData {
	lottieJsonPath: string
	singlePullSoundPath: string
	multiPullSoundPath: string
	mythicPullSoundPath: string
}

export interface LolTftTrovesTrovesCelebrationHighlightSegmentData {
	lottieJsonPath: string
	revealSoundPath: string
	transitionWipeSoundPath: string
	promiseTokenTitle: string
	promiseTokenDescription: string
}

export interface LolTftTrovesTrovesCelebrationPortalSegmentData {
	singlePullRareWebmPath: string
	singlePullEpicWebmPath: string
	singlePullLegendaryWebmPath: string
	singlePullMythicWebmPath: string
	singlePullSoundPath: string
	multiPullRareWebmPath: string
	multiPullEpicWebmPath: string
	multiPullLegendaryWebmPath: string
	multiPullMythicWebmPath: string
	multiPullSoundPath: string
}

export interface LolTftTrovesTrovesCelebrationStandardSegmentData {
	/** @format float */
	FirstItemTimingOffset: number
	/** @format float */
	InterItemTimingOffset: number
	PCHeaderText: string
	PCButtonText: string
	PCBackgroundPath: string
	PCRewardFramePath: string
	PCRewardOneStarPath: string
	PCRewardTwoStarPath: string
	PCRewardThreeStarPath: string
	PCRewardRareGemPath: string
	PCRewardEpicGemPath: string
	PCRewardLegendaryGemPath: string
	PCRewardMythicGemPath: string
	pullSingleIndividualGlintSoundPath: string
	pullSingleIndividualGlintLegendarySoundPath: string
	revealGlobalSoundPath: string
	revealEpicSoundPath: string
	revealMythicSoundPath: string
	revealRareSoundPath: string
	/** @format float */
	PCRewardFadeInDuration: number
	/** @format float */
	PCRewardFadeInDelay: number
	/** @format float */
	PCThumbnailFadeInDuration: number
	/** @format float */
	PCThumbnailFadeInDelay: number
	PCRewardSheenPath: string
	/** @format float */
	PCRewardSheenDuration: number
	/** @format float */
	PCRewardSheenDelay: number
	PCGlintSprite: LolTftTrovesTrovesPCSpriteAnimation
	PCLegendarySparkSprite: LolTftTrovesTrovesPCSpriteAnimation
	PCLegendaryHitSprite: LolTftTrovesTrovesPCSpriteAnimation
}

export interface LolTftTrovesTrovesCelebrationThemeData {
	currencySegmentData: LolTftTrovesTrovesCelebrationCurrencySegmentData
	portalSegmentData: LolTftTrovesTrovesCelebrationPortalSegmentData
	highlightSegmentData: LolTftTrovesTrovesCelebrationHighlightSegmentData
	standardSegmentData: LolTftTrovesTrovesCelebrationStandardSegmentData
}
// ...
export interface LolTftTrovesTrovesStatus {
	owned: boolean
	/** @format uint32 */
	availableContents: number
	/** @format uint8 */
	pityCount: number
	dropTableId: string
	hasPullError: boolean
	/** @format uint16 */
	totalRollsCount: number
	isCollectorBountyMaxRollsMet: boolean
}